refactored epochs
